授权

+++ app/Http/Controllers/Note/NoteController.php
@@ -10,17 +10,19 @@ use App\Http\Requests\Note\UpdateNoteRequest;
 use App\Library\Upload;
 use App\Models\Note;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class NoteController extends Controller
 {
     public function index(IndexNoteRequest $request)
     {
         $query = auth('api')->user()->note();
-        if ($request->filled('category_id')){
-            $query->where('category_id',$request->category_id);
+        if ($request->filled('category_id')) {
+            $query->where('category_id', $request->category_id);
         }
         $count = $query->count();
-        $data = $query->orderByDesc('id')->offset(($request->page - 1) * $request->limit)
+        $data = $query->orderByDesc('id')
+            ->offset(($request->page - 1) * $request->limit)
             ->limit($request->limit)
             ->get(['id', 'title', 'user_id', 'category_id', 'created_at']);
         return response()->json([
@@ -31,6 +33,7 @@ class NoteController extends Controller
 
     public function show(Request $request, Note $note)
     {
+        $this->authorize('view',$note->query()->find($request->note_id));
         $this->validate($request, [
             'note_id' => 'required'
         ]);
@@ -44,17 +47,19 @@ class NoteController extends Controller
         $note->category_id = $request->category_id;
         $note->user_id = auth('api')->user()->id;
         $note->save();
-        return response()->json(['message' => 'success'],201);
+        return response()->json(['message' => 'success'], 201);
     }
 
     public function update(UpdateNoteRequest $request, Note $note)
     {
+        $this->authorize('update',$note->query()->find($request->note_id));
         $note->query()->find($request->note_id)->update($request->except('note_id'));
         return response()->json(['message' => 'success']);
     }
 
     public function delete(DeleteNoteRequest $request, Note $note)
     {
+        $this->authorize('delete',$note->query()->find($request->note_id));
         $note->query()->find($request->note_id)->delete();
         return response()->json(['message' => 'success']);
     }
